Support browser-based auth (VSTS) in Eclipse 3.5.2
The target was updated to include the "swt" sub-folder,
in which SWT version 3.104.2 was copied from Eclipse 4.5.2.
To make sure SWT version 3.5.2 isn't used, those checkboxes were
unchecked in the "Content" pane, which had the effect of exploding all
the bundles into the <includeBundles> element.  I couldn't find an
exclusion mechanism!